+++ src/service.ts
@@ -3,9 +3,9 @@ import { pspec, registerGObject, PspecFlag, PspecType } from './utils/gobject.js
 
 export type OnlyString<S extends string | unknown> = S extends string ? S : never;
 
-export type Props<T> = Pick<T, {
+export type Props<T> = Omit<Pick<T, {
     [K in keyof T]: T[K] extends (...args: any[]) => any ? never : OnlyString<K>
-}[keyof T]>;
+}[keyof T]>, 'g_type_instance'>;
 
 export type BindableProps<T> = {
     [K in keyof T]: Binding<any, any, NonNullable<T[K]>> | T[K];
@@ -18,7 +18,7 @@ export class Binding<
 > {
     emitter: Emitter;
     prop: Prop;
-    transformFn = (v: Return) => v;
+    transformFn = (v: any) => v; // see #262
     constructor(emitter: Emitter, prop: Prop) {
         this.emitter = emitter;
         this.prop = prop;
@@ -27,7 +27,6 @@ export class Binding<
     transform<T>(fn: (v: Return) => T) {
         const bind = new Binding<Emitter, Prop, T>(this.emitter, this.prop);
         const prev = this.transformFn;
-        // @ts-expect-error
         bind.transformFn = (v: Return) => fn(prev(v));
         return bind;
     }